@@ -10,14 +10,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.1.1",
-    "@angular/common": "^19.1.1",
-    "@angular/compiler": "^19.1.1",
-    "@angular/core": "^19.1.1",
-    "@angular/forms": "^19.1.1",
-    "@angular/platform-browser": "^19.1.1",
-    "@angular/platform-browser-dynamic": "^19.1.1",
-    "@angular/router": "^19.1.1",
+    "@angular/animations": "^19.2.8",
+    "@angular/common": "^19.2.8",
+    "@angular/compiler": "^19.2.8",
+    "@angular/core": "^19.2.8",
+    "@angular/forms": "^19.2.8",
+    "@angular/platform-browser": "^19.2.8",
+    "@angular/platform-browser-dynamic": "^19.2.8",
+    "@angular/router": "^19.2.8",
     "bootstrap": "^5.3.3",
     "ngx-bootstrap": "^19.0.2",
     "rxjs": "~7.8.1",
@@ -25,9 +25,9 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.1.2",
-    "@angular/cli": "~19.1.2",
-    "@angular/compiler-cli": "^19.1.1",
+    "@angular-devkit/build-angular": "^19.2.9",
+    "@angular/cli": "~19.2.9",
+    "@angular/compiler-cli": "^19.2.8",
     "@types/jasmine": "~5.1.4",
     "angular-cli-ghpages": "^2.0.0-beta.2",
     "jasmine-core": "~5.1.2",
