@@ -1,8 +1,8 @@
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { Router } from '@angular/router';
-import { BehaviorSubject, Observable, throwError } from 'rxjs';
-import { catchError, map } from 'rxjs/operators';
+import { BehaviorSubject, Observable, of, throwError } from 'rxjs';
+import { catchError, delay, switchMap, tap } from 'rxjs/operators';
 import { User } from '../_interfaces/user';
 import { environment } from 'src/environments/environment.development';
 
@@ -10,29 +10,52 @@ import { environment } from 'src/environments/environment.development';
   providedIn: 'root'
 })
 export class AuthService {
+  private http = inject(HttpClient);
+  private router = inject(Router);
 
-  private userSubject: BehaviorSubject<User | null>;
-  public user$: Observable<User | null>;
+  private tokenSubject: BehaviorSubject<string | null> = new BehaviorSubject<string | null>(JSON.parse(localStorage.getItem('token')!));
+  public token$: Observable<string | null> = this.tokenSubject.asObservable();
   private apiUrl = environment.urlAddress;
 
-  constructor(private http: HttpClient, private router: Router) {
-    this.userSubject = new BehaviorSubject<User | null>(JSON.parse(localStorage.getItem('user')!));
-    this.user$ = this.userSubject.asObservable();
+  // Track login state with BehaviorSubject
+  private isUserLoggedInSubject = new BehaviorSubject<boolean>(false);
+
+  constructor() {
+    // Only initialize sessionStorage on the client-side (browser)
+    if (typeof window !== 'undefined' && window.localStorage) {
+      const storedLoginState = localStorage.getItem('isUserLoggedIn') === 'true';
+      this.isUserLoggedInSubject.next(storedLoginState);
+    }
   }
 
-  public get userValue(): User | null {
-    return this.userSubject.value;
+  public get tokenValue(): string | null {
+    return this.tokenSubject.value;
   }
 
-  login(email: string | null, password: string | null): Observable<User> {
+  // Expose the login status as an observable
+  get isUserLoggedIn$(): Observable<boolean> {
+    return this.isUserLoggedInSubject.asObservable();
+  }
+
+  login(email: string | null, password: string | null): Observable<boolean> {
     return this.http.post<any>(`${this.apiUrl}/customer/authenticate`, { email, password })
       .pipe(
-        map(user => {
+        switchMap((token) => {
+          const isLoggedIn = token !== null;
           //store customer detail and basic credentials in local storage
-          localStorage.setItem('user', JSON.stringify(user));
-          this.userSubject.next(user);
-          return user
+          if (typeof window !== 'undefined' && window.localStorage) {
+            localStorage.setItem('isUserLoggedIn', isLoggedIn ? 'true' : 'false');
+          }
+          // Update the BehaviorSubject with new login state
+          this.isUserLoggedInSubject.next(isLoggedIn);
+          localStorage.setItem('token', JSON.stringify(token));
+          this.tokenSubject.next(token);
+          return of(isLoggedIn).pipe(
+            delay(1000),
+            tap(val => console.log("Is User Authentication successful: " + val))
+          );
         }),
+
         catchError(err => {
           return throwError(() => new Error('Either Incorrect Username or Password'));
         })
@@ -48,21 +71,23 @@ export class AuthService {
 
   logout() {
     //log out user by removing from local storage
-    localStorage.removeItem('user');
-    this.userSubject.next(null);
+    localStorage.removeItem('token');
+    this.tokenSubject.next(null);
+
+    if (typeof window !== 'undefined' && window.localStorage) {
+      localStorage.removeItem('isUserLoggedIn');
+    }
+    // Update the BehaviorSubject to false when logged out
+    this.isUserLoggedInSubject.next(false);
     this.router.navigate(['/login']);
   }
 
-  getUser(): any {
-    return localStorage.getItem('user');
+  getToken(): any {
+    return localStorage.getItem('token');
   }
   isLoggedIn(): boolean {
-    return this.getUser() !== null;
-  }
-
-  canActivate() {
-
-  }
+    return this.getToken() !== null;
+  } 
 
   private handleError(error: HttpErrorResponse) {
     if (error.status === 0) {
