@@ -1,7 +1,9 @@
-import { Component, inject } from '@angular/core';
+import { Component, computed, effect, inject, signal, Signal } from '@angular/core';
 import { Product, ProductService } from './product.service';
 import { FormsModule } from '@angular/forms';
 import { CartService } from '../cart/cart.service';
+import { HttpErrorResponse } from '@angular/common/http';
+import { setErrorMessage } from '../shared/error-message';
 
 @Component({
   selector: 'app-product',
@@ -12,6 +14,7 @@ import { CartService } from '../cart/cart.service';
 export class ProductComponent {
   private productService = inject(ProductService);
   cartService = inject(CartService);
+  productId = signal<number>(0);
 
   products = this.productService.products;
   selectedProduct = this.productService.selectedProduct;
@@ -20,8 +23,17 @@ export class ProductComponent {
   color = this.productService.color;
   isLoading = this.productService.isLoading;
   errorMessage = this.productService.errorMessage;
+
+  productByIdResource = this.productService.getProductById(this.productId);
+  productById: Signal<Product> = computed(() => this.productByIdResource.value()?.data ?? {} as Product);
+  errorById: Signal<HttpErrorResponse> = computed(() => this.productByIdResource.error() as HttpErrorResponse);
+  errorMessageById = computed(() => setErrorMessage(this.errorById(), "ProductById"));
+  errorStatusById = computed(() => this.errorById().status)
+  isLoadingById: Signal<boolean> = this.productByIdResource.isLoading;
+
   addToCart(item: Product | undefined, qty: number) {
     window.alert('Your product has been added to the cart!');
     this.cartService.addToOrder(item, qty);
   }
+  productByIdEff = effect(() => console.log(this.selectedProduct()));
 }
